requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.6.1
drf-yasg==1.21.8
fastapi==0.112.0
frozenlist==1.5.0
gunicorn==21.2.0
h11==0.14.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
motor==3.5.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.10.1
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
sqlparse==0.5.3

# Phase 4: Real-Time Tracking
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.5
yarl==1.18.3
